+++ WebDiary.Frontend/Clients/diaryGroupClient.cs
@@ -1,37 +1,69 @@
 using System;
+using Microsoft.AspNetCore.Components.Authorization;
 using WebDiary.Frontend.Models;
+using WebDiary.Frontend.Models.Auth;
 
 namespace WebDiary.Frontend.Clients;
 
-public class DiaryGroupClient(HttpClient httpClient)
+public class DiaryGroupClient
 {
-    virtual public async Task<List<DiaryGroup>> GetGroupsAsync() =>
-        await httpClient.GetFromJsonAsync<List<DiaryGroup>>("groups") ?? new List<DiaryGroup>();
+    private readonly HttpClient httpClient;
+    private readonly CustomAuthenticationStateProvider? authStateProvider;
+
+    public DiaryGroupClient(HttpClient httpClient, AuthenticationStateProvider? authenticationStateProvider = null)
+    {
+        this.httpClient = httpClient;
+        authStateProvider = authenticationStateProvider as CustomAuthenticationStateProvider;
+    }
+
+    virtual public async Task<List<DiaryGroup>> GetGroupsAsync()
+    {
+        var response = await AuthorizedRequestAsync(() => httpClient.GetAsync("groups"));
+        return await response.Content.ReadFromJsonAsync<List<DiaryGroup>>() ?? new List<DiaryGroup>();
+    }
         
-    virtual public async Task<List<DiaryGroup>> GetGroupsOfUserAsync(int userId) =>
-        await httpClient.GetFromJsonAsync<List<DiaryGroup>>($"groups/ofuser/{userId}") ?? new List<DiaryGroup>();
+    virtual public async Task<List<DiaryGroup>> GetGroupsOfUserAsync(int userId)
+    {
+        var response = await AuthorizedRequestAsync(() => httpClient.GetAsync($"groups/ofuser/{userId}"));
+        return await response.Content.ReadFromJsonAsync<List<DiaryGroup>>() ?? new List<DiaryGroup>();
+    }
         
-    virtual public async Task<DiaryGroup> GetGroupAsync(int id) =>
-        await httpClient.GetFromJsonAsync<DiaryGroup>($"groups/{id}") ?? throw new Exception("Group wasn't found");
+    virtual public async Task<DiaryGroup> GetGroupAsync(int id)
+    {
+        var response = await AuthorizedRequestAsync(() => httpClient.GetAsync($"groups/{id}"));
+        return await response.Content.ReadFromJsonAsync<DiaryGroup>() ?? throw new Exception("Group wasn't found");
+    }
 
     public async Task AddGroupAsync(DiaryGroup group) {
-        var response = await httpClient.PostAsJsonAsync<DiaryGroup>("groups", group);
+        var response = await AuthorizedRequestAsync(() =>
+            httpClient.PostAsJsonAsync<DiaryGroup>("groups", group));
         if(!response.IsSuccessStatusCode) {
             throw new Exception();
         }
     }
 
     public async Task UpdateGroupAsync(DiaryGroup newGroup) {
-        var response = await httpClient.PutAsJsonAsync<DiaryGroup>($"groups/{newGroup.Id}", newGroup);
+        var response = await AuthorizedRequestAsync(() =>
+            httpClient.PutAsJsonAsync<DiaryGroup>($"groups/{newGroup.Id}", newGroup));
         if(!response.IsSuccessStatusCode) {
             throw new Exception();
         }
     }
 
     public async Task DeleteGroupAsync(int id) {
-        var response = await httpClient.DeleteAsync($"groups/{id}");
+        var response = await AuthorizedRequestAsync(() => httpClient.DeleteAsync($"groups/{id}"));
         if(!response.IsSuccessStatusCode) {
             throw new Exception();
         }
+    }
+
+    private async Task<HttpResponseMessage> AuthorizedRequestAsync(Func<Task<HttpResponseMessage>> action)
+    {
+        if (authStateProvider == null)
+        {
+            return await action();
+        }
+
+        return await authStateProvider.AuthorizedRequestAsync(action);
     }
 }
